feat(1-3267): some cleanup
